refactor(movement-optimizer): extract motion view scaffolding
