+++ src/movement_optimizer/gui/motion_tabs.py
@@ -3,7 +3,7 @@
 
 from __future__ import annotations
 
-from collections.abc import Callable, Sequence
+from collections.abc import Callable
 from itertools import pairwise
 from typing import cast
 
@@ -53,6 +53,7 @@ from movement_optimizer.rendering import Palette, get_chart_color
 from . import plot_renderer
 from .motion_analysis_panel import MotionAnalysisPanel
 from .motion_controls import NumericControl, scrollable_control_panel
+from .motion_view import MotionViewMixin
 from .policy_trace_canvas import PolicyTraceCanvas, refresh_policy_trace_palette
 from .policy_worker import PolicyOptimizationWorker
 from .vector_overlay import (
@@ -64,6 +65,8 @@ from .vector_overlay import (
     auto_scale_factor,
     draw_overlay_scene,
 )
+
+_MotionViewMixin = MotionViewMixin
 
 
 # Canvas colours are sourced from the fleet shared theme (via rendering.Palette
@@ -120,18 +123,12 @@ def _swing_overlay_scene(
     origin = (float(field.com_m[0]), float(field.com_m[1]))
     if gravity:
         gravity_vec = (float(field.gravity_n[0]), float(field.gravity_n[1]))
-        arrows.append(
-            ForceArrow(origin, gravity_vec, VectorStyle(LEG, label="gravity"))
-        )
+        arrows.append(ForceArrow(origin, gravity_vec, VectorStyle(LEG, label="gravity")))
     if tension:
         tension_vec = (float(field.chain_tension_n[0]), float(field.chain_tension_n[1]))
-        arrows.append(
-            ForceArrow(origin, tension_vec, VectorStyle(CHAIN, label="tension"))
-        )
+        arrows.append(ForceArrow(origin, tension_vec, VectorStyle(CHAIN, label="tension")))
     if torque:
-        for joint, magnitude in zip(
-            SWING_POLICY_JOINT_NAMES, field.joint_torque_nm, strict=True
-        ):
+        for joint, magnitude in zip(SWING_POLICY_JOINT_NAMES, field.joint_torque_nm, strict=True):
             point = field.joint_points_m[joint]
             arcs.append(
                 TorqueArc(
@@ -142,9 +139,7 @@ def _swing_overlay_scene(
             )
     if com:
         markers.append(ComMarker(origin, VectorStyle(ACCENT)))
-    return OverlayScene(
-        arrows=tuple(arrows), torque_arcs=tuple(arcs), com_markers=tuple(markers)
-    )
+    return OverlayScene(arrows=tuple(arrows), torque_arcs=tuple(arcs), com_markers=tuple(markers))
 
 
 def _chain_overlay_scene(
@@ -295,8 +290,7 @@ class MotionCanvas(QWidget):
     @staticmethod
     def _compute_chain_path_length(chain_nodes: list[tuple[float, float]]) -> float:
         distances = [
-            np.hypot(end[0] - start[0], end[1] - start[1])
-            for start, end in pairwise(chain_nodes)
+            np.hypot(end[0] - start[0], end[1] - start[1]) for start, end in pairwise(chain_nodes)
         ]
         return max(float(sum(distances)), 0.5)
 
@@ -342,96 +336,7 @@ class MotionCanvas(QWidget):
             )
 
 
-class _MotionViewMixin:
-    """Shared Animation/Plots subtab scaffolding for motion-analysis tabs.
-
-    Hosts the per-element animation layer toggles, the animation/plots
-    sub-tab split, and the plot-legend visibility control. Concrete tabs
-    must provide ``self.canvas`` (a :class:`MotionCanvas`),
-    ``self.analysis_panel`` (a :class:`MotionAnalysisPanel`), and the
-    ``self._layer_toggles`` / ``self._plot_legend_toggle`` attributes
-    referenced below.
-    """
-
-    canvas: MotionCanvas
-    analysis_panel: MotionAnalysisPanel
-    _layer_toggles: dict[str, QCheckBox]
-
-    def _build_animation_view(self) -> QWidget:
-        """The animation subtab: the motion canvas with full vertical room."""
-        view = QWidget()
-        view_layout = QVBoxLayout(view)
-        view_layout.setContentsMargins(0, 0, 0, 0)
-        view_layout.addWidget(self.canvas)
-        return view
-
-    def _build_plots_view(self) -> QWidget:
-        """The plots subtab: roomy analysis plots plus appearance controls."""
-        view = QWidget()
-        view_layout = QVBoxLayout(view)
-        view_layout.setContentsMargins(0, 0, 0, 0)
-        view_layout.setSpacing(6)
-        appearance = QHBoxLayout()
-        self._plot_legend_toggle = QCheckBox("Show plot legends")
-        self._plot_legend_toggle.setChecked(True)
-        self._plot_legend_toggle.setToolTip(
-            "Show or hide the legends on the analysis plots so they do not "
-            "obscure the plotted curves."
-        )
-        self._plot_legend_toggle.stateChanged.connect(self._refresh_plot_legends)
-        appearance.addWidget(self._plot_legend_toggle)
-        appearance.addStretch()
-        view_layout.addLayout(appearance)
-        plot_scroll = QScrollArea()
-        plot_scroll.setWidgetResizable(True)
-        plot_scroll.setFrameShape(QScrollArea.Shape.NoFrame)
-        plot_scroll.setWidget(self.analysis_panel)
-        view_layout.addWidget(plot_scroll, stretch=1)
-        return view
-
-    def _build_layers_group(self, layer_keys: Sequence[str] | None = None) -> QGroupBox:
-        """Build the "Show in animation" checklist.
-
-        ``layer_keys`` restricts the checklist to the layers a given tab
-        actually draws (e.g. the chain tab has no articulated rider), so no
-        inert toggles are shown. Defaults to every canvas layer.
-        """
-        allowed = set(layer_keys) if layer_keys is not None else None
-        group = QGroupBox("Show in animation")
-        layout = QVBoxLayout(group)
-        layout.setSpacing(4)
-        tips = {
-            "grid": "Background reference grid.",
-            "chain": "Swing chain polyline.",
-            "rider": "Articulated rider body segments.",
-            "markers": "Anchor and seat pivot markers.",
-            "forces": "All force and torque vector overlays.",
-        }
-        for key, label in MotionCanvas.LAYERS:
-            if allowed is not None and key not in allowed:
-                continue
-            checkbox = QCheckBox(label)
-            checkbox.setChecked(self.canvas.is_layer_visible(key))
-            checkbox.setToolTip(tips.get(key, ""))
-            checkbox.stateChanged.connect(
-                lambda _state, name=key, box=checkbox: self.canvas.set_layer_visible(
-                    name, box.isChecked()
-                )
-            )
-            self._layer_toggles[key] = checkbox
-            layout.addWidget(checkbox)
-        return group
-
-    def _apply_plot_legend_visibility(self) -> None:
-        """Match analysis-plot legend visibility to the appearance toggle."""
-        self.analysis_panel.set_legends_visible(self._plot_legend_toggle.isChecked())
-
-    def _refresh_plot_legends(self, _state: int | None = None) -> None:
-        self._apply_plot_legend_visibility()
-        self.analysis_panel.draw()
-
-
-class SwingsetTab(_MotionViewMixin, QWidget):
+class SwingsetTab(MotionViewMixin, QWidget):
     """Interactive swingset model tab with cyclic policy optimization."""
 
     playbackStateChanged = pyqtSignal()  # noqa: N815 - Qt signal naming convention.
@@ -739,30 +644,16 @@ class SwingsetTab(_MotionViewMixin, QWidget):
             integer=True,
             tooltip="Time steps simulated per evaluation when cycles are not used (up to 2000).",
         )
-        self._add_control(
-            form, "freq_min", "Freq min Hz", 0.2, 2.0, 0.45, refresh=False
-        )
-        self._add_control(
-            form, "freq_max", "Freq max Hz", 0.2, 2.0, 0.75, refresh=False
-        )
+        self._add_control(form, "freq_min", "Freq min Hz", 0.2, 2.0, 0.45, refresh=False)
+        self._add_control(form, "freq_max", "Freq max Hz", 0.2, 2.0, 0.75, refresh=False)
         self._add_control(
             form, "freq_samples", "Freq samples", 1, 8, 3, integer=True, refresh=False
         )
-        self._add_control(
-            form, "hip_rate_min", "Hip min rad/s", 0.0, 3.0, 0.5, refresh=False
-        )
-        self._add_control(
-            form, "hip_rate_max", "Hip max rad/s", 0.0, 3.0, 1.3, refresh=False
-        )
-        self._add_control(
-            form, "hip_samples", "Hip samples", 1, 8, 2, integer=True, refresh=False
-        )
-        self._add_control(
-            form, "torso_rate_min", "Torso min rad/s", 0.0, 3.0, 0.3, refresh=False
-        )
-        self._add_control(
-            form, "torso_rate_max", "Torso max rad/s", 0.0, 3.0, 1.1, refresh=False
-        )
+        self._add_control(form, "hip_rate_min", "Hip min rad/s", 0.0, 3.0, 0.5, refresh=False)
+        self._add_control(form, "hip_rate_max", "Hip max rad/s", 0.0, 3.0, 1.3, refresh=False)
+        self._add_control(form, "hip_samples", "Hip samples", 1, 8, 2, integer=True, refresh=False)
+        self._add_control(form, "torso_rate_min", "Torso min rad/s", 0.0, 3.0, 0.3, refresh=False)
+        self._add_control(form, "torso_rate_max", "Torso max rad/s", 0.0, 3.0, 1.1, refresh=False)
         self._add_control(
             form,
             "torso_samples",
@@ -773,12 +664,8 @@ class SwingsetTab(_MotionViewMixin, QWidget):
             integer=True,
             refresh=False,
         )
-        self._add_control(
-            form, "knee_ratio_min", "Knee ratio min", 0.0, 1.5, 0.25, refresh=False
-        )
-        self._add_control(
-            form, "knee_ratio_max", "Knee ratio max", 0.0, 1.5, 0.65, refresh=False
-        )
+        self._add_control(form, "knee_ratio_min", "Knee ratio min", 0.0, 1.5, 0.25, refresh=False)
+        self._add_control(form, "knee_ratio_max", "Knee ratio max", 0.0, 1.5, 0.65, refresh=False)
         self._add_control(
             form, "knee_samples", "Knee samples", 1, 8, 2, integer=True, refresh=False
         )
@@ -792,9 +679,7 @@ class SwingsetTab(_MotionViewMixin, QWidget):
             integer=True,
             refresh=False,
         )
-        self._add_control(
-            form, "speed", "Playback speed", 0.25, 4.0, 1.0, refresh=False
-        )
+        self._add_control(form, "speed", "Playback speed", 0.25, 4.0, 1.0, refresh=False)
         layout.addLayout(form)
         return group
 
@@ -1044,23 +929,15 @@ class SwingsetTab(_MotionViewMixin, QWidget):
     def _populate_analysis_panel(self) -> None:
         if self._rollout is None:
             return
-        history = swing_force_history(
-            self._config(), self._rollout, DEFAULT_POLICY_DT_S
-        )
+        history = swing_force_history(self._config(), self._rollout, DEFAULT_POLICY_DT_S)
         self._force_history = history
-        self._force_fields = swing_force_fields(
-            self._config(), self._rollout, DEFAULT_POLICY_DT_S
-        )
+        self._force_fields = swing_force_fields(self._config(), self._rollout, DEFAULT_POLICY_DT_S)
         panel = self.analysis_panel
         panel.clear()
-        plot_renderer.plot_swing_joint_torques(
-            panel.axes["torques"], history, legend=False
-        )
+        plot_renderer.plot_swing_joint_torques(panel.axes["torques"], history, legend=False)
         plot_renderer.plot_swing_joint_power(panel.axes["power"], history, legend=False)
         plot_renderer.plot_swing_angle(panel.axes["angle"], history, legend=False)
-        plot_renderer.plot_swing_com_height(
-            panel.axes["com_height"], history, legend=False
-        )
+        plot_renderer.plot_swing_com_height(panel.axes["com_height"], history, legend=False)
         plot_renderer.plot_swing_energy(panel.axes["energy"], history, legend=False)
         plot_renderer.plot_swing_com_path(panel.axes["com_path"], history, legend=False)
         self._apply_plot_legend_visibility()
